+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Catchy phrase here :)
 ## New features
 
 * Support for single element list dimension definition, which is
-considered to be of Category type
+considered to be of Category type. It can be used to fix the value 
+for some dimension during optimization.
 * Add iteration event handler for `BayesSearchCV`
 
 ## Version 0.4